+++ core/src/net/oschina/j2cache/MultiThreadTester.java
@@ -3,16 +3,17 @@ package net.oschina.j2cache;
 import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.TimeUnit;
 
 /**
  * 对 J2Cache 进行多线程测试
  */
 public class MultiThreadTester {
 
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) {
         CacheChannel cache = J2Cache.getChannel();
+
         ExecutorService threadPool = Executors.newCachedThreadPool();
+
         for(int i=0;i<100;i++) {
             final int seq = i;
             threadPool.execute(() -> {
@@ -26,9 +27,6 @@ public class MultiThreadTester {
                 }
             });
         }
-
-        threadPool.awaitTermination(Integer.MAX_VALUE, TimeUnit.MILLISECONDS);
-        System.exit(0);
     }
 
 }